feat(#4) - 강의 검색 기능 추가 및 TestCode 추가

+++ src/main/java/com/flab/infrun/lecture/infrastructure/persistence/mybatis/mapper/LectureMyBatisMapper.java
@@ -9,7 +9,7 @@ import org.apache.ibatis.annotations.Mapper;
 @Mapper
 public interface LectureMyBatisMapper {
 
-    int save(Lecture entity);
+    long save(Lecture entity);
 
     Optional<Lecture> findById(Long id);
