+++ src/main/java/com/flab/infrun/lecture/infrastructure/persistence/mybatis/mapper/LectureVideoFileMyBatisMapper.java
@@ -7,7 +7,7 @@ import org.apache.ibatis.annotations.Mapper;
 @Mapper
 public interface LectureVideoFileMyBatisMapper {
 
-    int save(LectureVideoFile entity);
+    long save(LectureVideoFile entity);
 
     Optional<LectureVideoFile> findById(Long id);
 }
